tests and readme
