+++ src/sparseml/transformers/finetune/__init__.py
@@ -17,7 +17,6 @@
 from .data import DataTrainingArguments, TextGenerationDataset
 from .model_args import ModelArguments
 from .session_mixin import SessionManagerMixIn
-from .sft_trainer import SFTTrainer
 from .text_generation import apply, compress, eval, oneshot, train
 from .trainer import Trainer
 from .training_args import TrainingArguments
